Profile page, area to show booked tickets
