+++ SP23.P03.Web/ClientApp/src/components/NavBar/NavBar.css
@@ -2,6 +2,6 @@
     background-color: #d8b4fe;
     margin: 10px 40px 10px 40px;
     border-radius: 30px;
-    height: 60px;
+    height: 65px;
 }
